Break up serialization and deserialization and make deserialization more flexible
